pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here - New"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "langchain>=1.2.10",
    "langchain-ollama>=0.2.0",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.1",
]

[tool.basedpyright]
venvPath = "."
venv = ".venv"
